:recycle: refactor: 에러 class utils/errors 폴더로 이동
