+++ frontend/src/utils/errors/AuthError.ts
