+++ frontend/src/utils/errors/GetError.ts
