+++ frontend/src/utils/errors/RequestError.ts
